Chore: Rework game definition
This makes it so that there's a class to represent a game since the data
structure continues to get a bit more complex and complex, plus it has
its own needs. This simplifies each game state and the MenuState, which
is nice.

I don't love how the definition lives in `Reg`, but I could see in the
future the definition living in a json file or something like that. Only
downside with a JSON file is that it doesn't have the nice type safety
of the Haxe definitions. So, maybe what happens is that it becomes a
haxe typedef and then loops through to instantiate. We'll see as more
and more games are added.